LayoutTests/crypto/crypto-random-values-expected.txt

Tests crypto.randomValues.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'crypto' in window is true
PASS 'getRandomValues' in window.crypto is true
PASS matchingBytes < 100 is true
PASS successfullyParsed is true

TEST COMPLETE
